마이페이지 회원정보 수정까지

+++ gym/src/main/java/com/gym/gym/mapper/AnswerMapper.java
@@ -11,17 +11,17 @@ public interface AnswerMapper {
 
 public List<Answer> list() throws Exception;
 
-public Answer select(int no) throws Exception;
+public Answer select(Long no) throws Exception;
 
     public int insert(Answer answer) throws Exception;
 
     public int update(Answer answer) throws Exception;
     
-    public int delete(int no) throws Exception;
+    public int delete(Long no) throws Exception;
 
-    public List<Answer> listByParent(int boardNo) throws Exception;
+    public List<Answer> listByParent(Long boardNo) throws Exception;
 
-    public int deleteByParent(int boardNo ) throws Exception;
+    public int deleteByParent(Long boardNo ) throws Exception;
